Refactored dcore_pre.py. Some functions were moved to interaction.py
